requirements.txt
langchain
python-dotenv
google-generativeai
serpapi
streamlit